% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_analysis.R
\name{analyze_time_trend}
\alias{analyze_time_trend}
\title{Analyze temporal trends in resistance genes}
\usage{
analyze_time_trend(data)
}
\arguments{
\item{data}{A data frame with resistome data}
}
\value{
A data frame with average abundance of each gene over time
}
\description{
Analyze temporal trends in resistance genes
}
