% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{forecast}
\alias{forecast}
\title{Create a forecast for abundance data by organism using ARIMA}
\usage{
forecast(data_time, gene_id, h = 10)
}
\arguments{
\item{data_time}{A filtered data frame with time trend data for one gene}

\item{gene_id}{The gene ID to forecast}

\item{h}{Number of future periods to forecast}
}
\value{
A data frame with mean, lower, and upper predictions for each organism
}
\description{
Create a forecast for abundance data by organism using ARIMA
}
