% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load}
\alias{load}
\title{Load resistome data with time and location metadata}
\usage{
load(file_path)
}
\arguments{
\item{file_path}{Path to the CSV file containing resistome data}
}
\value{
A data frame with resistome data
}
\description{
Load resistome data with time and location metadata
}
