% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{load_resistome_data}
\alias{load_resistome_data}
\title{Load resistome data with time and location metadata}
\usage{
load_resistome_data(file_path)
}
\arguments{
\item{file_path}{Path to the CSV file containing resistome data}
}
\value{
A data frame with resistome data
}
\description{
Load resistome data with time and location metadata
}
