% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_fc}
\alias{plot_fc}
\title{Plot temporal trends and forecast by organism}
\usage{
plot_fc(data_time, forecast_df, gene_id)
}
\arguments{
\item{data_time}{A data frame with time trend data}

\item{forecast_df}{The forecast data frame created by `forecast`}

\item{gene_id}{The gene ID to plot}
}
\value{
A ggplot object showing the temporal trend and forecast by organism
}
\description{
Plot temporal trends and forecast by organism
}
