% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_analysis.R
\name{plot_geo_distribution}
\alias{plot_geo_distribution}
\title{Plot geographical distribution of a specific resistance gene}
\usage{
plot_geo_distribution(data, gene_id)
}
\arguments{
\item{data}{A data frame with resistome data}

\item{gene_id}{The gene identifier to visualize}
}
\value{
A ggplot object showing the geographical distribution of the gene's abundance
}
\description{
Plot geographical distribution of a specific resistance gene
}
