% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_map}
\alias{plot_map}
\title{Plot geographical distribution for a specific gene}
\usage{
plot_map(data, gene_id)
}
\arguments{
\item{data}{A data frame with resistome data}

\item{gene_id}{The gene ID to plot}
}
\value{
A ggplot object with the geographical distribution
}
\description{
Plot geographical distribution for a specific gene
}
