% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_analysis.R
\name{plot_tr}
\alias{plot_tr}
\title{Plot temporal trends for a specific resistance gene and organism}
\usage{
plot_tr(data_time, gene_id, organism = NULL)
}
\arguments{
\item{data_time}{A data frame with time trend data (output of trend)}

\item{gene_id}{The gene ID to plot}

\item{organism}{(Optional) The organism to filter}
}
\value{
A ggplot object with the time trend
}
\description{
Plot temporal trends for a specific resistance gene and organism
}
