% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_analysis.R
\name{trend}
\alias{trend}
\title{Analyze temporal trends in resistance genes by organism}
\usage{
trend(data)
}
\arguments{
\item{data}{A data frame with resistome data}
}
\value{
A data frame with average abundance of each gene and organism over time
}
\description{
Analyze temporal trends in resistance genes by organism
}
